refactor: move common helpers into shared utils file
