+++ app/src/espresso/java/xyz/lbres/trickcalculator/ui/settings/settingsHelpers.kt
@@ -1,7 +1,6 @@
 package xyz.lbres.trickcalculator.ui.settings
 
 import androidx.test.espresso.Espresso.onView
-import androidx.test.espresso.action.ViewActions.click
 import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.matcher.ViewMatchers.isChecked
 import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
@@ -9,29 +8,6 @@ import androidx.test.espresso.matcher.ViewMatchers.isNotChecked
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import org.hamcrest.Matchers.allOf
 import xyz.lbres.trickcalculator.R
-import xyz.lbres.trickcalculator.testutils.viewactions.forceClick
-
-/**
- * Open settings fragment through attributions fragment
- */
-fun openFragment() {
-    val infoButton = onView(withId(R.id.infoButton))
-    infoButton.check(matches(isDisplayed()))
-    infoButton.perform(click())
-    onView(withId(R.id.title)).perform(click())
-}
-
-/**
- * Click close button
- */
-fun closeFragment() {
-    onView(withId(R.id.closeButton)).perform(forceClick())
-}
-
-/**
- * Wrapper function to create a [SettingsModifiedViewAssertion]
- */
-fun settingsRandomized() = SettingsModifiedViewAssertion()
 
 /**
  * Check that settings match initial settings.
