+++ app/src/main/java/xyz/lbres/trickcalculator/utils/ViewExt.kt
@@ -2,6 +2,8 @@ package xyz.lbres.trickcalculator.utils
 
 import android.view.View
 
+// TODO espresso tests for these functions
+
 /**
  * Enable view
  */
